新增了以加密形式保存Cookie至本地 新增用户组类,用于可同时操作不同已登入帐号

+++ cloudreve调用/Json/Admin/用户列表Json.cs
@@ -241,7 +241,7 @@ namespace cloudreve.Json.Admin
             /// <summary>
             /// 已使用空间大小
             /// </summary>
-            public int Storage { get; set; }
+            public long Storage { get; set; }
             /// <summary>
             /// 两步验证
             /// </summary>
